+++ CHANGES.rst
@@ -5,7 +5,9 @@ Changelog
 0.15 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Fix deprecated import AccessControl.class_init instead of App.class_init 
+  (Plone6 compatibility)
+  [boulch]
 
 
 0.14 (2021-06-29)